Update read_oo_pidata()
The function read data correctly but because of the missing information in the header, the read files could not be converted to irradiance. The function now allows the user to supply a whole instrument descriptor object, or alternatively to manually set some key fields in the constructed descriptor. In addition the integration time was wrong by a factor 10^6 because of a bug in the expected time unit.