Revise QC_dark()
limits are now expressed per 1000 pixels and the tolerance is relative to MAD. This should make the defaults equally good for the whole spectrum as for smaller bands. Using MAD, will hopefully take care of differences on the background and variability of the dark noise.

The calls in s_irrad_corrected() can now use the defaults built into QC_dark(). Further testing will show if this is true.